Updated code that was warned by compiler

+++ src/config/Config.c
@@ -26,20 +26,20 @@ void readTerminalArguments(OPTIONS * config, int argc, char ** argv) {
   if (argc < ARGC_MIN) printCriticalError(ERROR_BAD_ARGUMENTS, "Arguments not specified");
 
   config->filepath = argv[MEDIAPATH_ARGV_INDEX];
-  for (size_t i = MEDIAPATH_ARGV_INDEX + 1; i < argc; ++i) {
-    if (strcmp(argv[i], OPTION_FPS) == 0 && i + 1 < argc) 
+  for (size_t i = MEDIAPATH_ARGV_INDEX + 1; (int)i < argc; ++i) {
+    if (strcmp(argv[i], OPTION_FPS) == 0 && (int)(i + 1) < argc) 
     {
       int FPS = atoi(argv[++i]);
       if (FPS < MIN_FPS) printCriticalError(ERROR_BAD_ARGUMENTS, "FPS is smaller than 0 (unlimited)");
       config->FPS = (size_t)FPS;
     } 
-    else if (strcmp(argv[i], OPTION_BLOCKSIZE) == 0 && i + 1 < argc) 
+    else if (strcmp(argv[i], OPTION_BLOCKSIZE) == 0 && (int)(i + 1) < argc) 
     {
       int blockSize = atoi(argv[++i]);
       if (blockSize < MIN_BLOCKSIZE) printCriticalError(ERROR_BAD_ARGUMENTS, "blockSize is smaller than 1"); 
       config->blockSize = (size_t)blockSize;
     }
-    else if (strcmp(argv[i], OPTION_OUTPUT) == 0 && i + 1 < argc) 
+    else if (strcmp(argv[i], OPTION_OUTPUT) == 0 && (int)(i + 1) < argc) 
     {
       config->output = argv[++i];
     }
